+++ lib/features/feat_posts/models/post_model.dart
@@ -17,5 +17,7 @@ class PostModel {
 
   String get image => record.getStringValue('image');
 
+  String get category => record.getStringValue('category');
+
   RecordModel toRecord() => record;
 }